# devoir-api-port-russell/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "API Port Russel",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/users/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "slice(7",
            "in": "header",
            "type": "string"
          },
          {
            "name": "length",
            "in": "header",
            "type": "string"
          },
          {
            "name": "w-access-token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/users/dashboard": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "slice(7",
            "in": "header",
            "type": "string"
          },
          {
            "name": "length",
            "in": "header",
            "type": "string"
          },
          {
            "name": "w-access-token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/users/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "slice(7",
            "in": "header",
            "type": "string"
          },
          {
            "name": "length",
            "in": "header",
            "type": "string"
          },
          {
            "name": "w-access-token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "501": {
            "description": "Not Implemented"
          }
        }
      },
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "slice(7",
            "in": "header",
            "type": "string"
          },
          {
            "name": "length",
            "in": "header",
            "type": "string"
          },
          {
            "name": "w-access-token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "firstname": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "501": {
            "description": "Not Implemented"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "slice(7",
            "in": "header",
            "type": "string"
          },
          {
            "name": "length",
            "in": "header",
            "type": "string"
          },
          {
            "name": "w-access-token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "501": {
            "description": "Not Implemented"
          }
        }
      }
    },
    "/catways/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "slice(7",
            "in": "header",
            "type": "string"
          },
          {
            "name": "length",
            "in": "header",
            "type": "string"
          },
          {
            "name": "w-access-token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/reservations/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "slice(7",
            "in": "header",
            "type": "string"
          },
          {
            "name": "length",
            "in": "header",
            "type": "string"
          },
          {
            "name": "w-access-token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/dashboard": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "slice(7",
            "in": "header",
            "type": "string"
          },
          {
            "name": "length",
            "in": "header",
            "type": "string"
          },
          {
            "name": "w-access-token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }
  }
}